CDAP-15590 do not copy if artifact exists

+++ cdap-app-fabric/src/main/java/io/cdap/cdap/internal/app/runtime/plugin/PluginInstantiator.java
@@ -131,7 +131,9 @@ public class PluginInstantiator implements Closeable {
    */
   public void addArtifact(Location artifactLocation, ArtifactId destArtifact) throws IOException {
     File destFile = new File(pluginDir, Artifacts.getFileName(destArtifact));
-    Files.copy(Locations.newInputSupplier(artifactLocation), destFile);
+    if (!destFile.exists()) {
+      Files.copy(Locations.newInputSupplier(artifactLocation), destFile);
+    }
   }
 
   /**
